Création page d'erreur authentification et création système authentification

+++ models/entrances/delete.php
@@ -1,26 +1,32 @@
 <?php
 
-if(isset($_GET['id']) && !empty($_GET['id'])) {
-    require_once('../core/config.php');
+session_start();
 
-    $id = $_GET['id'];
+if($_SESSION['user']['role'] === 'role_admin') {
+    if(isset($_GET['id']) && !empty($_GET['id'])) {
+        require_once('../core/config.php');
 
-    $check = $pdo->prepare("SELECT * FROM entrances WHERE id = :id");
-    $check->bindValue(':id', $id, PDO::PARAM_STR);
-    $check->execute(); 
-    $checkEntrance = $check->fetch();
+        $id = $_GET['id'];
 
-    if(!$checkEntrance) {
-        $messageError = "Cette entrée n'existe pas";
-    }
+        $check = $pdo->prepare("SELECT * FROM entrances WHERE id = :id");
+        $check->bindValue(':id', $id, PDO::PARAM_STR);
+        $check->execute(); 
+        $checkEntrance = $check->fetch();
 
-    $delete = $pdo->prepare('DELETE FROM entrances WHERE id = :id');
-    $delete->bindValue(':id', $checkEntrance['id'], PDO::PARAM_STR);
-    if ($delete->execute()) {
-        $messageSuccess = "Produit supprimé";
+        if(!$checkEntrance) {
+            $messageError = "Cette entrée n'existe pas";
+        }
 
-        header('Location: /entrees');
+        $delete = $pdo->prepare('DELETE FROM entrances WHERE id = :id');
+        $delete->bindValue(':id', $checkEntrance['id'], PDO::PARAM_STR);
+        if ($delete->execute()) {
+            $messageSuccess = "Produit supprimé";
+
+            header('Location: /entrees');
+        }
+    } else {
+        header('Location: /404');
     }
 } else {
-    header('Location: /404');
+    header('Location: /non-autorisee');
 }
